#193: move WindowDropHandler to entry-point package, illustrative ADR example
WindowDropHandler lived in transfer/ (data tier) but imports
presentation.RootComponent — a layering violation. Move it to
com.tubetoast.tether (entry-point tier, same package as MainUi) which
may import presentation.

ADR example now uses <RootScreenComposable> placeholder instead of a
fake concrete symbol, tying to the two-composable presentation rule.
